NTT-199 | Restructure directories and bump dependencies
