--- app/uk/gov/hmrc/trustsindividualcheck/controllers/MicroserviceHelloWorldController.scala
@@ -1,17 +1,0 @@
-package uk.gov.hmrc.trustsindividualcheck.controllers
-
-import javax.inject.{Inject, Singleton}
-import play.api.mvc.{Action, AnyContent, ControllerComponents}
-import uk.gov.hmrc.play.bootstrap.backend.controller.BackendController
-import uk.gov.hmrc.trustsindividualcheck.config.AppConfig
-
-import scala.concurrent.Future
-
-@Singleton()
-class MicroserviceHelloWorldController @Inject()(appConfig: AppConfig, cc: ControllerComponents)
-    extends BackendController(cc) {
-
-  def hello(): Action[AnyContent] = Action.async { implicit request =>
-    Future.successful(Ok("Hello world"))
-  }
-}
