+++ conf/application.conf
@@ -1,38 +1,33 @@
+# Copyright 2020 HM Revenue & Customs
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 
 include "backend.conf"
 
-# This is the main configuration file for the application.
-# ~~~~~
-
 appName = trusts-individual-check
 
-# An ApplicationLoader that uses Guice to bootstrap the application.
 play.application.loader = "uk.gov.hmrc.play.bootstrap.ApplicationLoader"
 
-# Primary entry point for all HTTP requests on Play applications
 play.http.requestHandler = "uk.gov.hmrc.play.bootstrap.http.RequestHandler"
 
-# Provides an implementation of AuditConnector. Use `uk.gov.hmrc.play.bootstrap.AuditModule` or create your own.
-# An audit connector must be provided.
 play.modules.enabled += "uk.gov.hmrc.play.bootstrap.AuditModule"
-
-# Provides an implementation of MetricsFilter. Use `uk.gov.hmrc.play.graphite.GraphiteMetricsModule` or create your own.
-# A metric filter must be provided
 play.modules.enabled += "uk.gov.hmrc.play.bootstrap.graphite.GraphiteMetricsModule"
-
-# Provides an implementation and configures all filters required by a Platform frontend microservice.
 play.modules.enabled += "uk.gov.hmrc.play.bootstrap.backend.BackendModule"
-play.http.filters = "uk.gov.hmrc.play.bootstrap.backend.filters.BackendFilters"
-
-# Default http client
 play.modules.enabled += "uk.gov.hmrc.play.bootstrap.HttpClientModule"
 
-# Json error handler
-play.http.errorHandler = "uk.gov.hmrc.play.bootstrap.backend.http.JsonErrorHandler"
+play.http.filters = "uk.gov.hmrc.play.bootstrap.filters.MicroserviceFilters"
 
-# Play Modules
-# ~~~~
-# Additional play modules can be added here
+play.http.errorHandler = "uk.gov.hmrc.play.bootstrap.backend.http.JsonErrorHandler"
 
 
 # Secret key
@@ -40,9 +35,9 @@ play.http.errorHandler = "uk.gov.hmrc.play.bootstrap.backend.http.JsonErrorHandl
 # The secret key is used to secure cryptographics functions.
 # If you deploy your application to several instances be sure to use the same key!
 play.http.secret.key = "8DKbKxnjhIhx845kJVE29F9eEEL2QgOOlnEs3dzajIZZYFZzjKR1YoV6d8PHhze7"
+application.session.httpOnly = false
+application.session.secure = false
 
-# The application languages
-# ~~~~~
 play.i18n.langs = ["en"]
 
 # Router
@@ -57,31 +52,17 @@ play.i18n.langs = ["en"]
 # !!!WARNING!!! DO NOT CHANGE THIS ROUTER
 play.http.router = prod.Routes
 
-# Controller
-# ~~~~~
-# By default all controllers will have authorisation, logging and
-# auditing (transaction monitoring) enabled.
-# The below controllers are the default exceptions to this rule.
+logger.root = ERROR
+logger.play = INFO
+logger.application = DEBUG
 
-controllers {
-  # 300 is the default, you may need to change this according to your needs
-  confidenceLevel = 300
-
-  uk.gov.hmrc.trustsindividualcheck.controllers.MicroserviceHelloWorldController = {
-    needsAuth = false
-    needsLogging = false
-    needsAuditing = false
-  }
-}
-
-# Metrics plugin settings - graphite reporting is configured on a per env basis
 metrics {
   name = ${appName}
   rateUnit = SECONDS
   durationUnit = SECONDS
   showSamples = true
   jvm = true
-  enabled = true
+  enabled = false
 }
 
 # Microservice specific config
