+++ project/AppDependencies.scala
@@ -6,14 +6,31 @@ import sbt._
 object AppDependencies {
 
   val compile = Seq(
-    "uk.gov.hmrc"             %% "bootstrap-backend-play-27"  % "2.23.0"
+    play.sbt.PlayImport.ws,
+    "uk.gov.hmrc"             %% "bootstrap-backend-play-27"        % "2.23.0",
+    "org.reactivemongo"       %% "play2-reactivemongo"              % "0.20.9-play27"
   )
 
   val test = Seq(
-    "uk.gov.hmrc"             %% "bootstrap-test-play-27"   % "2.23.0" % Test,
-    "org.scalatest"           %% "scalatest"                % "3.1.2"                 % Test,
-    "com.typesafe.play"       %% "play-test"                % current                 % Test,
-    "com.vladsch.flexmark"    %  "flexmark-all"             % "0.35.10"               % "test, it",
-    "org.scalatestplus.play"  %% "scalatestplus-play"       % "4.0.3"                 % "test, it"
+    "uk.gov.hmrc"             %% "bootstrap-test-play-27"   % "2.23.0",
+    "org.scalatest"           %% "scalatest"                % "3.1.2",
+    "com.typesafe.play"       %% "play-test"                % current,
+    "com.vladsch.flexmark"    %  "flexmark-all"             % "0.35.10",
+    "org.scalatestplus.play"  %% "scalatestplus-play"       % "4.0.3",
+    "org.mockito"             %  "mockito-all"              % "1.10.19",
+    "org.scalacheck"          %% "scalacheck"               % "1.14.1"
+  ).map(_ % "test,it")
+
+  def apply(): Seq[ModuleID] = compile ++ test
+
+  val akkaVersion = "2.5.23"
+  val akkaHttpVersion = "10.0.15"
+
+  val overrides = Seq(
+    "com.typesafe.akka" %% "akka-stream" % akkaVersion,
+    "com.typesafe.akka" %% "akka-protobuf" % akkaVersion,
+    "com.typesafe.akka" %% "akka-slf4j" % akkaVersion,
+    "com.typesafe.akka" %% "akka-actor" % akkaVersion,
+    "com.typesafe.akka" %% "akka-http-core" % akkaHttpVersion
   )
 }
